requirements.txt
discord
paho-mqtt
